backend/requirements.txt
langgraph
langchain
langchain-community
langserve
langsmith
langchain_groq
dotenv
aiofiles
streamlit
graphviz
transformers
torch
pillow
pymupdf
pymupdf4llm
pdfplumber
pdfminer.six
numpy 
pandas
spire.pdf
ollama_ocr